backend/auth_server/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
httpx==0.28.*
PyJWT==2.10.*
google-cloud-secret-manager==2.*
simple-salesforce==1.*
starlette==0.41.*
